Code cleanup + review of Field model

+++ src/sdk/PnP.Core.Test/SharePoint/ListFieldTests.cs
@@ -1125,7 +1125,7 @@ namespace PnP.Core.Test.SharePoint
                 Assert.AreEqual("TEST GROUP", addedField.Group);
                 Assert.AreEqual(FieldType.Choice, addedField.FieldTypeKind);
                 Assert.IsTrue(addedField.FillInChoice);
-                Assert.AreEqual((int)ChoiceFormatType.RadioButtons, addedField.EditFormat);
+                Assert.AreEqual(ChoiceFormatType.RadioButtons, addedField.EditFormat);
                 // Currently supports collections only for model types (JsonMappingHelper Ln 120)
                 Assert.IsNotNull(addedField.Choices);
                 Assert.AreEqual(2, addedField.Choices.Length);
@@ -1157,7 +1157,7 @@ namespace PnP.Core.Test.SharePoint
                 Assert.AreEqual("TEST GROUP", addedField.Group);
                 Assert.AreEqual(FieldType.Choice, addedField.FieldTypeKind);
                 Assert.IsTrue(addedField.FillInChoice);
-                Assert.AreEqual((int)ChoiceFormatType.RadioButtons, addedField.EditFormat);
+                Assert.AreEqual(ChoiceFormatType.RadioButtons, addedField.EditFormat);
                 // Currently supports collections only for model types (JsonMappingHelper Ln 120)
                 Assert.IsNotNull(addedField.Choices);
                 Assert.AreEqual(2, addedField.Choices.Length);
@@ -1190,7 +1190,7 @@ namespace PnP.Core.Test.SharePoint
                 Assert.AreEqual("TEST GROUP", addedField.Group);
                 Assert.AreEqual(FieldType.Choice, addedField.FieldTypeKind);
                 Assert.IsTrue(addedField.FillInChoice);
-                Assert.AreEqual((int)ChoiceFormatType.RadioButtons, addedField.EditFormat);
+                Assert.AreEqual(ChoiceFormatType.RadioButtons, addedField.EditFormat);
                 // Currently supports collections only for model types (JsonMappingHelper Ln 120)
                 Assert.IsNotNull(addedField.Choices);
                 Assert.AreEqual(2, addedField.Choices.Length);
@@ -1223,7 +1223,7 @@ namespace PnP.Core.Test.SharePoint
                 Assert.AreEqual("TEST GROUP", addedField.Group);
                 Assert.AreEqual(FieldType.Choice, addedField.FieldTypeKind);
                 Assert.IsTrue(addedField.FillInChoice);
-                Assert.AreEqual((int)ChoiceFormatType.RadioButtons, addedField.EditFormat);
+                Assert.AreEqual(ChoiceFormatType.RadioButtons, addedField.EditFormat);
                 // Currently supports collections only for model types (JsonMappingHelper Ln 120)
                 Assert.IsNotNull(addedField.Choices);
                 Assert.AreEqual(2, addedField.Choices.Length);
@@ -1257,7 +1257,7 @@ namespace PnP.Core.Test.SharePoint
                 Assert.AreEqual("TEST GROUP", addedField.Group);
                 Assert.AreEqual(FieldType.Choice, addedField.FieldTypeKind);
                 Assert.IsTrue(addedField.FillInChoice);
-                Assert.AreEqual((int)ChoiceFormatType.RadioButtons, addedField.EditFormat);
+                Assert.AreEqual(ChoiceFormatType.RadioButtons, addedField.EditFormat);
                 // Currently supports collections only for model types (JsonMappingHelper Ln 120)
                 Assert.IsNotNull(addedField.Choices);
                 Assert.AreEqual(2, addedField.Choices.Length);
@@ -1291,7 +1291,7 @@ namespace PnP.Core.Test.SharePoint
                 Assert.AreEqual("TEST GROUP", addedField.Group);
                 Assert.AreEqual(FieldType.Choice, addedField.FieldTypeKind);
                 Assert.IsTrue(addedField.FillInChoice);
-                Assert.AreEqual((int)ChoiceFormatType.RadioButtons, addedField.EditFormat);
+                Assert.AreEqual(ChoiceFormatType.RadioButtons, addedField.EditFormat);
                 // Currently supports collections only for model types (JsonMappingHelper Ln 120)
                 Assert.IsNotNull(addedField.Choices);
                 Assert.AreEqual(2, addedField.Choices.Length);
